...actually, the previous commit was a mistake.
I realized it was a mistake before even committing it, but
I'm weird and like to keep a history of what I've done even
if it was wrong.

It turns out DST is accounted for. The problem was actually
the test, which expected the gmt_offset to always be the
CURRENT local GMT offset. In fact it should be the GMT
offset for the given date (which it is).

This reverts commit 6a6a90c04d5227ffcbda8863a88da229b7e1e4ac.