always provide the option to check for libyaml vulnerability
It occurred to me that this could happen: the user's OS has a patched
version of libyaml, BUT he/she is using a version of Psych with a
vendored libyaml that is still vulnerable (e.g., via Bundler). In this
case the check using dpkg -s libyaml-0-2 would incorrectly say the
system is safe! So really the only way to be sure is to run the check.